backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend con FastAPI y uv"
authors = [{ name = "Tu Nombre" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
]

[tool.uv]
# uv creará y manejará automáticamente un entorno virtual.
# Puedes agregar dependencias con "uv add"

[tool.uv.scripts]
dev = "uvicorn src.main:app --host 0.0.0.0 --port 8000 --reload"
start = "uvicorn src.main:app --host 0.0.0.0 --port 8000"
